station

+++ dotNet5782_1974_0044/PL/Hundlers/StationHundler.cs
@@ -72,7 +72,7 @@ namespace PL
             {
                 Id = (int)station.Id,
                 Name = station.Name,
-                AvailableChargingPorts = station.EmptyChargeSlots,
+                AvailableChargingPorts = (int)station.EmptyChargeSlots,
                 Location = LocationHandler.ConvertBackLocation(station.Location)
 
             };
